initial physics commit and debugging window addition

+++ control.h
@@ -6,4 +6,4 @@ void controlInit();
  * Return a boolean solid buffer representing new walls differenced
  * into the scene since calibration
  */
-int* threshhold(IplImage* calibration, float near, float far);
+uint8_t* threshhold(IplImage* calibration, float near, float far);
